requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
arabic-reshaper==3.0.0
argostranslate==1.11.0
asgiref==3.11.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
cloudpathlib==0.23.0
coloredlogs==15.0.1
confection==0.1.5
ctranslate2==4.7.1
cymem==2.0.13
Django==4.2.30
django-browser-reload==1.21.0
django-tailwind==3.6.0
emoji==2.15.0
filelock==3.19.1
flatbuffers==25.12.19
fsspec==2025.10.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
langdetect==1.0.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
minisbd==0.9.5
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.2.1
numpy==2.0.2
onnxruntime==1.19.2
packaging==26.1
pdf2image==1.17.0
pillow==11.3.0
preshed==3.0.13
protobuf==6.33.6
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
pymupdf==1.24.14
python-bidi==0.4.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==15.0.0
sacremoses==0.1.1
sentencepiece==0.2.1
shellingham==1.5.4
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.3
stanza==1.10.1
sympy==1.14.0
thinc==8.3.9
tomli==2.4.1
torch==2.8.0
tqdm==4.67.3
typer==0.23.2
typer-slim==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.2
